title and nav work
Added a better title that overrides the default title for each method in the News.php controller. Also created a makeLinks() function in our common_helper.php that uses data from the custom_config.php file. Changes to the header.php in the theme are required to use the makeLinks() function.